+++ src/sll/file.c
@@ -22,34 +22,37 @@ __SLL_EXTERNAL sll_file_t* sll_stderr=&_file_stderr;
 
 
 void _file_init_std_streams(void){
-	*((sll_file_descriptor_t*)(&(_file_stdin.fd)))=_platform_get_stream_descriptor(0);
-	sll_string_from_pointer(SLL_CHAR("sll_stdin"),(sll_string_t*)(&(_file_stdin.nm)));
+	*((sll_file_descriptor_t*)(&(_file_stdin.dt.fl.fd)))=_platform_get_stream_descriptor(0);
+	sll_string_from_pointer(SLL_CHAR("sll_stdin"),(sll_string_t*)(&(_file_stdin.dt.fl.nm)));
 	*((sll_file_flags_t*)(&(_file_stdin.f)))=SLL_FILE_FLAG_READ|SLL_FILE_FLAG_NO_BUFFER;
 	_file_stdin._l_num=0;
 	_file_stdin._l_off=0;
 	_file_stdin._off=0;
-	*((sll_file_descriptor_t*)(&(_file_stdout.fd)))=_platform_get_stream_descriptor(1);
-	sll_string_from_pointer(SLL_CHAR("sll_stdout"),(sll_string_t*)(&(_file_stdout.nm)));
+	*((sll_file_descriptor_t*)(&(_file_stdout.dt.fl.fd)))=_platform_get_stream_descriptor(1);
+	sll_string_from_pointer(SLL_CHAR("sll_stdout"),(sll_string_t*)(&(_file_stdout.dt.fl.nm)));
 	*((sll_file_flags_t*)(&(_file_stdout.f)))=SLL_FILE_FLAG_WRITE|SLL_FILE_FLAG_NO_BUFFER;
-	*((sll_file_descriptor_t*)(&(_file_stderr.fd)))=_platform_get_stream_descriptor(2);
-	sll_string_from_pointer(SLL_CHAR("sll_stderr"),(sll_string_t*)(&(_file_stderr.nm)));
+	*((sll_file_descriptor_t*)(&(_file_stderr.dt.fl.fd)))=_platform_get_stream_descriptor(2);
+	sll_string_from_pointer(SLL_CHAR("sll_stderr"),(sll_string_t*)(&(_file_stderr.dt.fl.nm)));
 	*((sll_file_flags_t*)(&(_file_stderr.f)))=SLL_FILE_FLAG_WRITE|SLL_FILE_FLAG_NO_BUFFER;
 }
 
 
 
 void _file_release_std_streams(void){
-	sll_deinit_string((sll_string_t*)(&(sll_stdin->nm)));
-	sll_deinit_string((sll_string_t*)(&(sll_stdout->nm)));
-	sll_deinit_string((sll_string_t*)(&(sll_stderr->nm)));
+	sll_deinit_string((sll_string_t*)(&(sll_stdin->dt.fl.nm)));
+	sll_deinit_string((sll_string_t*)(&(sll_stdout->dt.fl.nm)));
+	sll_deinit_string((sll_string_t*)(&(sll_stderr->dt.fl.nm)));
 }
 
 
 
 __SLL_EXTERNAL void sll_file_close(sll_file_t* f){
 	sll_file_flush(f);
-	sll_platform_file_close(f->fd);
-	sll_deinit_string((sll_string_t*)(&(f->nm)));
+	if (!(f->f&FILE_FLAG_BUFFER)){
+		sll_platform_file_close(f->dt.fl.fd);
+		sll_deinit_string((sll_string_t*)(&(f->dt.fl.nm)));
+	}
+	*((sll_file_flags_t*)(&(f->f)))=0;
 }
 
 
@@ -63,12 +66,18 @@ __SLL_EXTERNAL void sll_file_flush(sll_file_t* f){
 
 
 __SLL_EXTERNAL void sll_file_from_data(void* p,sll_size_t sz,sll_file_flags_t f,sll_file_t* o){
-	SLL_UNIMPLEMENTED();
+	*((void**)(&(o->dt.mm.p)))=p;
+	*((sll_size_t*)(&(o->dt.mm.sz)))=sz;
+	*((sll_file_flags_t*)(&(o->f)))=f|FILE_FLAG_BUFFER;
+	o->_l_num=0;
+	o->_l_off=0;
+	o->_off=0;
 }
 
 
 
 __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_file_open(const sll_char_t* nm,sll_file_flags_t f,sll_file_t* o){
+	SLL_ASSERT(!(f&FILE_FLAG_BUFFER));
 	if ((f&SLL_FILE_FLAG_READ)&&!sll_platform_path_exists(nm)){
 		return 0;
 	}
@@ -76,14 +85,12 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_bool_t sll_file_open(const sll_char_t* nm,
 	if (fd==SLL_UNKNOWN_FILE_DESCRIPTOR){
 		return 0;
 	}
-	*((sll_file_descriptor_t*)(&(o->fd)))=fd;
-	sll_string_from_pointer(nm,(sll_string_t*)(&(o->nm)));
+	*((sll_file_descriptor_t*)(&(o->dt.fl.fd)))=fd;
+	sll_string_from_pointer(nm,(sll_string_t*)(&(o->dt.fl.nm)));
 	*((sll_file_flags_t*)(&(o->f)))=f;
-	if (f&SLL_FILE_FLAG_READ){
-		o->_l_num=0;
-		o->_l_off=0;
-		o->_off=0;
-	}
+	o->_l_num=0;
+	o->_l_off=0;
+	o->_off=0;
 	return 1;
 }
 
@@ -93,17 +100,29 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_size_t sll_file_read(sll_file_t* f,void* p
 	if (!(f->f&SLL_FILE_FLAG_READ)){
 		return 0;
 	}
-	return sll_platform_file_read(f->fd,p,sz);
+	if (f->f&FILE_FLAG_BUFFER){
+		if (f->_off+sz>=f->dt.mm.sz){
+			sz=f->dt.mm.sz-f->_off;
+		}
+		sll_copy_data((void*)((uint64_t)(f->dt.mm.p)+f->_off),sz,p);
+		f->_off+=sz;
+		return sz;
+	}
+	sz=sll_platform_file_read(f->dt.fl.fd,p,sz);
+	f->_off+=sz;
+	return sz;
 }
 
 
 
 __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_read_char_t sll_file_read_char(sll_file_t* f){
+	if (!(f->f&SLL_FILE_FLAG_READ)){
+		return SLL_END_OF_DATA;
+	}
 	sll_char_t o;
 	if (!sll_file_read(f,&o,sizeof(sll_char_t))){
 		return SLL_END_OF_DATA;
 	}
-	f->_off++;
 	if (o=='\n'){
 		f->_l_num++;
 		f->_l_off=f->_off;
@@ -114,7 +133,9 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_read_char_t sll_file_read_char(sll_file_t*
 
 
 __SLL_EXTERNAL void sll_file_reset(sll_file_t* f){
-	sll_platform_file_seek(f->fd,0);
+	if (!(f->f&FILE_FLAG_BUFFER)){
+		sll_platform_file_seek(f->dt.fl.fd,0);
+	}
 	f->_l_num=0;
 	f->_l_off=0;
 	f->_off=0;
@@ -123,9 +144,29 @@ __SLL_EXTERNAL void sll_file_reset(sll_file_t* f){
 
 
 __SLL_EXTERNAL void sll_file_reset_line(sll_file_t* f,sll_file_offset_t off){
-	sll_platform_file_seek(f->fd,off);
+	if (!(f->f&SLL_FILE_FLAG_READ)){
+		return;
+	}
+	if (f->f&FILE_FLAG_BUFFER){
+		if (off>=f->dt.mm.sz){
+			f->_l_num=0;
+			f->_l_off=0;
+			f->_off=f->dt.mm.sz;
+			return;
+		}
+		const sll_char_t* dt=f->dt.mm.p;
+		while (off&&dt[off]!='\n'&&dt[off]!='\r'){
+			off--;
+		}
+		if (dt[off]=='\n'||dt[off]=='\r'){
+			off++;
+		}
+		f->_off=off;
+		return;
+	}
+	sll_platform_file_seek(f->dt.fl.fd,off);
 	sll_char_t c;
-	if (!sll_platform_file_read(f->fd,&c,sizeof(sll_char_t))){
+	if (!sll_file_read(f->dt.fl.fd,&c,sizeof(sll_char_t))){
 		f->_l_num=0;
 		f->_l_off=0;
 		f->_off=0;
@@ -133,8 +174,8 @@ __SLL_EXTERNAL void sll_file_reset_line(sll_file_t* f,sll_file_offset_t off){
 	}
 	while (off&&c!='\n'&&c!='\r'){
 		off--;
-		sll_platform_file_seek(f->fd,off);
-		if (!sll_platform_file_read(f->fd,&c,sizeof(sll_char_t))){
+		sll_platform_file_seek(f->dt.fl.fd,off);
+		if (!sll_platform_file_read(f->dt.fl.fd,&c,sizeof(sll_char_t))){
 			f->_l_num=0;
 			f->_l_off=0;
 			f->_off=0;
@@ -142,7 +183,7 @@ __SLL_EXTERNAL void sll_file_reset_line(sll_file_t* f,sll_file_offset_t off){
 		}
 	}
 	if (c!='\n'&&c!='\r'){
-		sll_platform_file_seek(f->fd,off);
+		sll_platform_file_seek(f->dt.fl.fd,off);
 	}
 	else{
 		off++;
@@ -156,11 +197,30 @@ __SLL_EXTERNAL sll_size_t sll_file_write(sll_file_t* f,const void* p,sll_size_t 
 	if (!(f->f&SLL_FILE_FLAG_WRITE)){
 		return 0;
 	}
-	return sll_platform_file_write(f->fd,p,sz);
+	if (f->f&FILE_FLAG_BUFFER){
+		if (f->_off+sz>=f->dt.mm.sz){
+			sz=f->dt.mm.sz-f->_off;
+		}
+		sll_copy_data(p,sz,((sll_char_t*)(f->dt.mm.p))+f->_off);
+		f->_off+=sz;
+		return sz;
+	}
+	return sll_platform_file_write(f->dt.fl.fd,p,sz);
 }
 
 
 
-__SLL_EXTERNAL void sll_file_write_char(sll_file_t* f,sll_char_t c){
-	sll_platform_file_write(f->fd,&c,sizeof(sll_char_t));
+__SLL_EXTERNAL sll_bool_t sll_file_write_char(sll_file_t* f,sll_char_t c){
+	if (!(f->f&SLL_FILE_FLAG_WRITE)){
+		return 0;
+	}
+	if (f->f&FILE_FLAG_BUFFER){
+		if (f->_off+1>=f->dt.mm.sz){
+			return 0;
+		}
+		*(((sll_char_t*)(f->dt.mm.p))+f->_off)=c;
+		f->_off++;
+		return 1;
+	}
+	return sll_platform_file_write(f->dt.fl.fd,&c,sizeof(sll_char_t));
 }